fix bitmap text install bug
