+++ src/scene/text/TextView.ts
@@ -1,20 +1,17 @@
-import { Cache } from '../../assets/cache/Cache';
 import { ObservablePoint } from '../../maths/point/ObservablePoint';
 import { emptyViewObserver } from '../../rendering/renderers/shared/view/View';
 import { uid } from '../../utils/data/uid';
-import { BitmapFont } from './bitmap/BitmapFont';
 import { BitmapFontManager } from './bitmap/BitmapFontManager';
-import { DynamicBitmapFont } from './bitmap/DynamicBitmapFont';
 import { CanvasTextMetrics } from './canvas/CanvasTextMetrics';
-import { HTMLTextStyle } from './html/HtmlTextStyle';
 import { measureHtmlText } from './html/utils/measureHtmlText';
+import { detectRenderType } from './utils/detectRenderType';
 import { ensureTextStyle } from './utils/ensureTextStyle';
 
 import type { PointData } from '../../maths/point/PointData';
 import type { View, ViewObserver } from '../../rendering/renderers/shared/view/View';
 import type { Bounds } from '../container/bounds/Bounds';
 import type { TextureDestroyOptions, TypeOrBool } from '../container/destroyTypes';
-import type { HTMLTextStyleOptions } from './html/HtmlTextStyle';
+import type { HTMLTextStyle, HTMLTextStyleOptions } from './html/HtmlTextStyle';
 import type { TextStyle, TextStyleOptions } from './TextStyle';
 
 export type TextString = string | number | { toString: () => string };
@@ -66,7 +63,7 @@ export class TextView implements View
     {
         this.text = options.text ?? '';
 
-        const renderMode = options.renderMode ?? this._detectRenderType(options.style);
+        const renderMode = options.renderMode ?? detectRenderType(options.style);
 
         this._renderMode = renderMode;
 
@@ -209,23 +206,6 @@ export class TextView implements View
             bounds[2] = (-anchor._y * height) - padding;
             bounds[3] = bounds[2] + height;
         }
-    }
-
-    private _detectRenderType(style: TextStyleOptions | AnyTextStyle): 'canvas' | 'html' | 'bitmap'
-    {
-        if (style instanceof HTMLTextStyle)
-        {
-            return 'html';
-        }
-
-        const fontData = Cache.get(`${style?.fontFamily as string}-bitmap`);
-
-        if (fontData instanceof DynamicBitmapFont || fontData instanceof BitmapFont)
-        {
-            return 'bitmap';
-        }
-
-        return 'canvas';
     }
 
     /**
